requirements.txt
openai
SpeechRecognition
pyaudio
keyboard
python-dotenv
pygame
